requirement.txt
gradio
ultralytics
torch
torchvision
torchaudio
Pillow
numpy
opencv-python
matplotlib
scipy
pandas
